Clean

+++ alias.zshrc
@@ -2,11 +2,13 @@ alias ~="cd ~/"
 alias ..="cd ../"
 alias ...="cd ../../"
 
+
 alias python=python3
 alias pip=pip3
-alias k=kubectl
-alias tf=terraform
+# alias k=kubectl
+# alias tf=terraform
 
-alias cip="docker inspect -f '{{.NetworkSettings.IPAddress}}'"
-alias crm="docker rm -f $(docker ps -aq)"
-alias cri="docker rmi $(docker images -q)"
+
+[ -f var/run/docker.sock ] && alias cip="docker inspect -f '{{.NetworkSettings.IPAddress}}'"
+[ -f var/run/docker.sock ] && alias crm="docker rm -f $(docker ps -aq)"
+[ -f var/run/docker.sock ] && alias cri="docker rmi $(docker images -q)"
